+++ airbyte-integrations/connectors/source-harvest/unit_tests/config.py
@@ -7,13 +7,27 @@ from typing import Any, Dict
 class ConfigBuilder:
     def __init__(self) -> None:
         self._config: Dict[str, Any] = {
-            "account_id": "any account id",
-            "credentials": {"api_token": "any api token"},
-            "replication_start_date": "2017-01-25T00:00:00Z",
+            "account_id": "an account id",
+            "replication_start_date": "2021-01-01T00:00:00Z",
+            "credentials": {
+                "api_token": "an api key"
+            }
         }
 
+    def with_account_id(self, account_id: str) -> "ConfigBuilder":
+        self._config["account_id"] = account_id
+        return self
+
     def with_replication_start_date(self, replication_start_date: datetime) -> "ConfigBuilder":
-        self._config["replication_start_date"] = replication_start_date.isoformat()
+        self._config["start_date"] = replication_start_date.isoformat()[:-13]+"Z"
+        return self
+
+    def with_api_token(self, api_token: str) -> "ConfigBuilder":
+        self._config["credentials"]["api_token"] = api_token
+        return self
+
+    def with_client_id(self, client_id: str) -> "ConfigBuilder":
+        self._config["credentials"]["client_id"] = client_id
         return self
 
     def build(self) -> Dict[str, Any]:
