+++ airbyte-integrations/connectors/source-harvest/unit_tests/integration/test_invoice_messages.py
@@ -1,24 +1,22 @@
 # Copyright (c) 2023 Airbyte, Inc., all rights reserved.
 
-from datetime import datetime, timedelta, timezone
+from datetime import datetime
 from typing import Any, Dict, Optional
 from unittest import TestCase
 
-from airbyte_cdk.sources.source import TState
 from airbyte_cdk.test.catalog_builder import CatalogBuilder
 from airbyte_cdk.test.entrypoint_wrapper import EntrypointOutput, read
-from airbyte_cdk.test.mock_http import HttpMocker, HttpRequest, HttpResponse
+from airbyte_cdk.test.mock_http import HttpMocker, HttpRequest
 from airbyte_cdk.test.mock_http.response_builder import (
     FieldPath,
     HttpResponseBuilder,
-    NestedPath,
     RecordBuilder,
     create_record_builder,
     create_response_builder,
     find_template,
 )
-from airbyte_protocol.models import ConfiguredAirbyteCatalog, FailureType, SyncMode
-from integration.config import ConfigBuilder
+from airbyte_protocol.models import SyncMode
+from config import ConfigBuilder
 from source_harvest import SourceHarvest
 
 _A_REPLICATION_START_DATE = "2021-01-01T00:00:00+00:00"
