+++ airbyte-integrations/connectors/source-harvest/unit_tests/test_source.py
@@ -1,5 +1,3 @@
-# Copyright (c) 2023 Airbyte, Inc., all rights reserved.
-
 import unittest
 from unittest.mock import Mock, patch
 
@@ -8,9 +6,10 @@ import requests
 from airbyte_cdk import AirbyteLogger
 from airbyte_cdk.models import FailureType
 from airbyte_cdk.utils import AirbyteTracedException
-from config import ConfigBuilder
 from requests import HTTPError
+
 from source_harvest.source import SourceHarvest
+from config import ConfigBuilder
 
 
 def _a_response(status_code: int) -> requests.Response:
@@ -27,6 +26,23 @@ class SourceTest(unittest.TestCase):
         self._source = SourceHarvest()
         self._logger = Mock(spec=AirbyteLogger)
         self._config = ConfigBuilder().build()
+
+    def test_given_config_with_client_id_without_account_id_when_check_connection_then_raise_config_error(self) -> None:
+        config = ConfigBuilder().with_client_id("a client id").build()
+        config.pop("account_id")
+
+        with pytest.raises(AirbyteTracedException) as exception_trace:
+            self._source.check_connection(self._logger, config)
+        assert exception_trace.value.failure_type == FailureType.config_error
+
+    def test_given_config_no_authentication_in_config_when_check_connection_then_raise_config_error(self) -> None:
+        config = ConfigBuilder().build()
+        config["credentials"].pop("api_token", None)
+        config["credentials"].pop("client_id", None)
+
+        with pytest.raises(AirbyteTracedException) as exception_trace:
+            self._source.check_connection(self._logger, config)
+        assert exception_trace.value.failure_type == FailureType.config_error
 
     @patch("source_harvest.source.Users.read_records")
     def test_given_400_http_error_when_check_connection_then_raise_non_config_error(self, mocked_user_read_records) -> None:
